+++ fieldservice_sale/README.rst
@@ -25,7 +25,8 @@ Field Service - Sales
 
 |badge1| |badge2| |badge3| |badge4| |badge5| 
 
-The module integrate the field service application with the sales one.
+The module integrates the Field Service application with the Sales one and
+allows you to sell field service products.
 
 **Table of contents**
 
@@ -50,10 +51,9 @@ To setup a product for quoting and selling field service orders:
 * Set the Product Type to 'Service' under General Information tab
 * Under Invoicing tab, set the Field Service Tracking option
 * Select the FSM Order Template that will be used for creating FSM Orders when
-    a Sale Order is confirmed with this product
+  a Sale Order is confirmed with this product
 
-
-To setup a sales territory, you need to:
+To setup a sales territory:
 
 * Go to Field Service > Master Data > Locations
 * Create or select a location
@@ -62,12 +62,12 @@ To setup a sales territory, you need to:
 Usage
 =====
 
-* Go to Sales app
+* Go to Sales
 * Create a new Quotation/Sale Order
 * Set the FSM Location to be used
 * On a Sale Order Line, select a product configured for field service orders
 * Confirm the Sale Order
-* Field Service Orders linked to SO lines are created
+* Field Service orders linked to SO lines are created
 
 Known issues / Roadmap
 ======================
@@ -96,19 +96,15 @@ Authors
 Contributors
 ~~~~~~~~~~~~
 
-* Steve Campbell <scampbell@opensourceintegrators.com>
-* Maxime Chambreuil <mchambreuil@opensourceintegrators.com>
-* Wolfgang Hall <whall@opensourceintegrators.com>
+* Open Source Integrators <https://opensourceintegrators.com>
+
+  * Steve Campbell <scampbell@opensourceintegrators.com>
+  * Maxime Chambreuil <mchambreuil@opensourceintegrators.com>
+  * Wolfgang Hall <whall@opensourceintegrators.com>
+
 * Serpent Consulting Services Pvt. Ltd. <support@serpentcs.com>
 * Brian McMaster <brian@mcmpest.com>
 * Rapha??l Reverdy <raphael.reverdy@akretion.com>
-
-Other credits
-~~~~~~~~~~~~~
-
-The development of this module has been financially supported by:
-
-* Open Source Integrators <https://opensourceintegrators.com>
 
 Maintainers
 ~~~~~~~~~~~
