+++ fieldservice_sale/__manifest__.py
@@ -1,23 +1,21 @@
-# Copyright (C) 2018 - TODAY, Open Source Integrators
+# Copyright (C) 2018 Open Source Integrators
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 {
     'name': 'Field Service - Sales',
-    'version': '12.0.1.0.1',
-    'summary': 'Sales integration for Field Service',
+    'version': '12.0.1.1.0',
+    'summary': 'Sell field services.',
     'category': 'Field Service',
     'author': 'Open Source Integrators, Odoo Community Association (OCA)',
     'website': 'https://github.com/OCA/field-service',
     'depends': [
         'fieldservice',
-        'fieldservice_recurring',
         'sale_management',
     ],
     'data': [
         'views/fsm_location.xml',
         'views/fsm_order.xml',
-        'views/fsm_recurring.xml',
-        'views/product.xml',
+        'views/product_template.xml',
         'views/sale_order.xml',
     ],
     'license': 'AGPL-3',
