+++ fieldservice_sale/i18n/fieldservice_sale.pot
@@ -1,0 +1,203 @@
+# Translation of Odoo Server.
+# This file contains the translation of the following modules:
+#	* fieldservice_sale
+#
+msgid ""
+msgstr ""
+"Project-Id-Version: Odoo Server 12.0\n"
+"Report-Msgid-Bugs-To: \n"
+"Last-Translator: <>\n"
+"Language-Team: \n"
+"MIME-Version: 1.0\n"
+"Content-Type: text/plain; charset=UTF-8\n"
+"Content-Transfer-Encoding: \n"
+"Plural-Forms: \n"
+
+#. module: fieldservice_sale
+#: model_terms:ir.ui.view,arch_db:fieldservice_sale.view_fsm_order_form_inherit_sale
+msgid "<span class=\"o_stat_text\">Sale Orders</span>"
+msgstr ""
+
+#. module: fieldservice_sale
+#: selection:product.template,field_service_tracking:0
+msgid "Create a recurring order"
+msgstr ""
+
+#. module: fieldservice_sale
+#: selection:product.template,field_service_tracking:0
+msgid "Create a single order"
+msgstr ""
+
+#. module: fieldservice_sale
+#: selection:product.template,field_service_tracking:0
+msgid "Don't create order"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model_terms:ir.ui.view,arch_db:fieldservice_sale.view_order_form_inherit_sale_fieldservice
+msgid "FSM Orders"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model_terms:ir.ui.view,arch_db:fieldservice_sale.view_order_form_inherit_sale_fieldservice
+msgid "Field Service Information"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model,name:fieldservice_sale.model_fsm_location
+msgid "Field Service Location"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model,name:fieldservice_sale.model_fsm_order
+msgid "Field Service Order"
+msgstr ""
+
+#. module: fieldservice_sale
+#: code:addons/fieldservice_sale/models/sale_order_line.py:67
+#, python-format
+msgid "Field Service Order Created (%s): <a href=\n"
+"                   # data-oe-model=fsm.order data-oe-id=%d>%s</a>\n"
+"                "
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_product_product__fsm_order_template_id
+#: model:ir.model.fields,field_description:fieldservice_sale.field_product_template__fsm_order_template_id
+msgid "Field Service Order Template"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,help:fieldservice_sale.field_sale_order_line__fsm_order_id
+msgid "Field Service Order generated by the sales order item"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_sale_order__fsm_order_count
+#: selection:product.template,service_type:0
+msgid "Field Service Orders"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_product_product__field_service_tracking
+#: model:ir.model.fields,field_description:fieldservice_sale.field_product_template__field_service_tracking
+msgid "Field Service Tracking"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_sale_order__fsm_order_ids
+msgid "Field Service orders associated to this sale"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_sale_order_line__is_field_service
+msgid "Is a Field Service"
+msgstr ""
+
+#. module: fieldservice_sale
+#: selection:product.template,service_type:0
+msgid "Manually set quantities on order"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,help:fieldservice_sale.field_product_product__service_type
+#: model:ir.model.fields,help:fieldservice_sale.field_product_template__service_type
+msgid "Manually set quantities on order: Invoice based on the manually entered quantity, without creating an analytic account.\n"
+"Timesheets on contract: Invoice based on the tracked hours on the related timesheet.\n"
+"Create a task and track hours: Create a task on the sales order validation and track the work hours."
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,help:fieldservice_sale.field_product_product__field_service_tracking
+#: model:ir.model.fields,help:fieldservice_sale.field_product_template__field_service_tracking
+msgid "On Sales order confirmation, this product can generate a field\n"
+"                service order or field service recurring order."
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_sale_order_line__fsm_order_id
+msgid "Order"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model,name:fieldservice_sale.model_product_template
+msgid "Product Template"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_sale_order__date_fsm_request
+msgid "Requested Service Date"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,help:fieldservice_sale.field_sale_order__fsm_location_id
+msgid "SO Lines generating a FSM order will be for this location"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_fsm_order__sale_line_id
+msgid "Sale Line"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model,name:fieldservice_sale.model_sale_order
+msgid "Sale Order"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model_terms:ir.ui.view,arch_db:fieldservice_sale.fsm_location_form_view
+msgid "Sales"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model,name:fieldservice_sale.model_sale_order_line
+msgid "Sales Order Line"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,help:fieldservice_sale.field_sale_order_line__is_field_service
+msgid "Sales Order item should generate an Order and/or\n"
+"                Order Recurrence, depending on the product settings."
+msgstr ""
+
+#. module: fieldservice_sale
+#: code:addons/fieldservice_sale/models/fsm_order.py:21
+#, python-format
+msgid "Sales Orders"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_fsm_location__sales_territory_id
+msgid "Sales Territory"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,help:fieldservice_sale.field_product_product__fsm_order_template_id
+#: model:ir.model.fields,help:fieldservice_sale.field_product_template__fsm_order_template_id
+msgid "Select the field service order template to be created"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_sale_order__fsm_location_id
+msgid "Service Location"
+msgstr ""
+
+#. module: fieldservice_sale
+#: code:addons/fieldservice_sale/models/sale_order_line.py:73
+#, python-format
+msgid "This order has been created from: <a href=\n"
+"                   # data-oe-model=sale.order data-oe-id=%d>%s</a> (%s)\n"
+"                "
+msgstr ""
+
+#. module: fieldservice_sale
+#: selection:product.template,service_type:0
+msgid "Timesheets on project (one fare per SO/Project)"
+msgstr ""
+
+#. module: fieldservice_sale
+#: model:ir.model.fields,field_description:fieldservice_sale.field_product_product__service_type
+#: model:ir.model.fields,field_description:fieldservice_sale.field_product_template__service_type
+msgid "Track Service"
+msgstr ""
+
