+++ fieldservice_sale/readme/USAGE.rst
@@ -1,6 +1,6 @@
-* Go to Sales app
+* Go to Sales
 * Create a new Quotation/Sale Order
 * Set the FSM Location to be used
 * On a Sale Order Line, select a product configured for field service orders
 * Confirm the Sale Order
-* Field Service Orders linked to SO lines are created
+* Field Service orders linked to SO lines are created
