+++ fieldservice_sale/static/description/index.html
@@ -368,7 +368,8 @@ ul.auto-toc {
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
 <p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/field-service/tree/12.0/fieldservice_sale"><img alt="OCA/field-service" src="https://img.shields.io/badge/github-OCA%2Ffield--service-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/field-service-12-0/field-service-12-0-fieldservice_sale"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runbot.odoo-community.org/runbot/264/12.0"><img alt="Try me on Runbot" src="https://img.shields.io/badge/runbot-Try%20me-875A7B.png" /></a></p>
-<p>The module integrate the field service application with the sales one.</p>
+<p>The module integrates the Field Service application with the Sales one and
+allows you to sell field service products.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
@@ -380,8 +381,7 @@ ul.auto-toc {
 <li><a class="reference internal" href="#credits" id="id6">Credits</a><ul>
 <li><a class="reference internal" href="#authors" id="id7">Authors</a></li>
 <li><a class="reference internal" href="#contributors" id="id8">Contributors</a></li>
-<li><a class="reference internal" href="#other-credits" id="id9">Other credits</a></li>
-<li><a class="reference internal" href="#maintainers" id="id10">Maintainers</a></li>
+<li><a class="reference internal" href="#maintainers" id="id9">Maintainers</a></li>
 </ul>
 </li>
 </ul>
@@ -400,13 +400,10 @@ ul.auto-toc {
 <li>Create or select a product</li>
 <li>Set the Product Type to ???Service??? under General Information tab</li>
 <li>Under Invoicing tab, set the Field Service Tracking option</li>
-<li><dl class="first docutils">
-<dt>Select the FSM Order Template that will be used for creating FSM Orders when</dt>
-<dd>a Sale Order is confirmed with this product</dd>
-</dl>
-</li>
+<li>Select the FSM Order Template that will be used for creating FSM Orders when
+a Sale Order is confirmed with this product</li>
 </ul>
-<p>To setup a sales territory, you need to:</p>
+<p>To setup a sales territory:</p>
 <ul class="simple">
 <li>Go to Field Service &gt; Master Data &gt; Locations</li>
 <li>Create or select a location</li>
@@ -416,12 +413,12 @@ ul.auto-toc {
 <div class="section" id="usage">
 <h1><a class="toc-backref" href="#id3">Usage</a></h1>
 <ul class="simple">
-<li>Go to Sales app</li>
+<li>Go to Sales</li>
 <li>Create a new Quotation/Sale Order</li>
 <li>Set the FSM Location to be used</li>
 <li>On a Sale Order Line, select a product configured for field service orders</li>
 <li>Confirm the Sale Order</li>
-<li>Field Service Orders linked to SO lines are created</li>
+<li>Field Service orders linked to SO lines are created</li>
 </ul>
 </div>
 <div class="section" id="known-issues-roadmap">
@@ -448,23 +445,19 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <div class="section" id="contributors">
 <h2><a class="toc-backref" href="#id8">Contributors</a></h2>
 <ul class="simple">
+<li>Open Source Integrators &lt;<a class="reference external" href="https://opensourceintegrators.com">https://opensourceintegrators.com</a>&gt;<ul>
 <li>Steve Campbell &lt;<a class="reference external" href="mailto:scampbell&#64;opensourceintegrators.com">scampbell&#64;opensourceintegrators.com</a>&gt;</li>
 <li>Maxime Chambreuil &lt;<a class="reference external" href="mailto:mchambreuil&#64;opensourceintegrators.com">mchambreuil&#64;opensourceintegrators.com</a>&gt;</li>
 <li>Wolfgang Hall &lt;<a class="reference external" href="mailto:whall&#64;opensourceintegrators.com">whall&#64;opensourceintegrators.com</a>&gt;</li>
+</ul>
+</li>
 <li>Serpent Consulting Services Pvt. Ltd. &lt;<a class="reference external" href="mailto:support&#64;serpentcs.com">support&#64;serpentcs.com</a>&gt;</li>
 <li>Brian McMaster &lt;<a class="reference external" href="mailto:brian&#64;mcmpest.com">brian&#64;mcmpest.com</a>&gt;</li>
 <li>Rapha??l Reverdy &lt;<a class="reference external" href="mailto:raphael.reverdy&#64;akretion.com">raphael.reverdy&#64;akretion.com</a>&gt;</li>
 </ul>
 </div>
-<div class="section" id="other-credits">
-<h2><a class="toc-backref" href="#id9">Other credits</a></h2>
-<p>The development of this module has been financially supported by:</p>
-<ul class="simple">
-<li>Open Source Integrators &lt;<a class="reference external" href="https://opensourceintegrators.com">https://opensourceintegrators.com</a>&gt;</li>
-</ul>
-</div>
 <div class="section" id="maintainers">
-<h2><a class="toc-backref" href="#id10">Maintainers</a></h2>
+<h2><a class="toc-backref" href="#id9">Maintainers</a></h2>
 <p>This module is maintained by the OCA.</p>
 <a class="reference external image-reference" href="https://odoo-community.org"><img alt="Odoo Community Association" src="https://odoo-community.org/logo.png" /></a>
 <p>OCA, or the Odoo Community Association, is a nonprofit organization whose
