+++ fieldservice_sale/views/fsm_order.xml
@@ -1,4 +1,3 @@
-<?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="view_fsm_order_form_inherit_sale" model="ir.ui.view">
         <field name="model">fsm.order</field>
